+++ src/Hpack/Config.hs
@@ -153,13 +153,15 @@ instance FromJSON CustomSetupSection where
 data LibrarySection = LibrarySection {
   librarySectionExposed :: Maybe Bool
 , librarySectionExposedModules :: Maybe (List String)
+, librarySectionGeneratedExposedModules :: Maybe (List String)
 , librarySectionOtherModules :: Maybe (List String)
+, librarySectionGeneratedOtherModules :: Maybe (List String)
 , librarySectionReexportedModules :: Maybe (List String)
 , librarySectionSignatures :: Maybe (List String)
 } deriving (Eq, Show, Generic)
 
 emptyLibrarySection :: LibrarySection
-emptyLibrarySection = LibrarySection Nothing Nothing Nothing Nothing Nothing
+emptyLibrarySection = LibrarySection Nothing Nothing Nothing Nothing Nothing Nothing Nothing
 
 instance HasFieldNames LibrarySection
 
@@ -169,10 +171,11 @@ instance FromJSON LibrarySection where
 data ExecutableSection = ExecutableSection {
   executableSectionMain :: Maybe FilePath
 , executableSectionOtherModules :: Maybe (List String)
+, executableSectionGeneratedOtherModules :: Maybe (List String)
 } deriving (Eq, Show, Generic)
 
 emptyExecutableSection :: ExecutableSection
-emptyExecutableSection = ExecutableSection Nothing Nothing
+emptyExecutableSection = ExecutableSection Nothing Nothing Nothing
 
 instance HasFieldNames ExecutableSection
 
@@ -557,6 +560,7 @@ data Library = Library {
   libraryExposed :: Maybe Bool
 , libraryExposedModules :: [String]
 , libraryOtherModules :: [String]
+, libraryGeneratedModules :: [String]
 , libraryReexportedModules :: [String]
 , librarySignatures :: [String]
 } deriving (Eq, Show)
@@ -564,6 +568,7 @@ data Library = Library {
 data Executable = Executable {
   executableMain :: Maybe FilePath
 , executableOtherModules :: [String]
+, executableGeneratedModules :: [String]
 } deriving (Eq, Show)
 
 data Section a = Section {
@@ -909,32 +914,34 @@ toLibrary dir name globalOptions =
     getLibraryModules Library{..} = libraryExposedModules ++ libraryOtherModules
 
     fromLibrarySectionTopLevel pathsModule inferableModules LibrarySection{..} =
-      Library librarySectionExposed exposedModules otherModules reexportedModules signatures
+      Library librarySectionExposed exposedModules otherModules generatedModules reexportedModules signatures
       where
-        (exposedModules, otherModules) =
-          determineModules pathsModule inferableModules librarySectionExposedModules librarySectionOtherModules
+        (exposedModules, otherModules, generatedModules) =
+          determineModules pathsModule inferableModules librarySectionExposedModules librarySectionGeneratedExposedModules librarySectionOtherModules librarySectionGeneratedOtherModules
         reexportedModules = fromMaybeList librarySectionReexportedModules
         signatures = fromMaybeList librarySectionSignatures
 
-determineModules :: [String] -> [String] -> Maybe (List String) -> Maybe (List String) -> ([String], [String])
-determineModules pathsModule inferableModules mExposedModules mOtherModules = case (mExposedModules, mOtherModules) of
-  (Nothing, Nothing) -> (inferableModules, pathsModule)
-  _ -> (exposedModules, otherModules)
-    where
-      exposedModules = maybe (inferableModules \\ otherModules) fromList mExposedModules
-      otherModules   = maybe ((inferableModules ++ pathsModule) \\ exposedModules) fromList mOtherModules
+determineModules :: [String] -> [String] -> Maybe (List String) -> Maybe (List String) -> Maybe (List String) -> Maybe (List String) -> ([String], [String], [String])
+determineModules pathsModule inferableModules mExposedModules mGeneratedExposedModules mOtherModules mGeneratedOtherModules = case (mExposedModules, mOtherModules) of
+  (Nothing, Nothing) -> (inferableModules ++ maybe [] fromList mGeneratedExposedModules, pathsModule ++ maybe [] fromList mGeneratedOtherModules, generatedModules)
+  _ -> (exposedModules, otherModules, generatedModules)
+  where
+    exposedModules = maybe (inferableModules \\ otherModules) fromList mExposedModules ++ maybe [] fromList mGeneratedExposedModules
+    otherModules   = maybe ((inferableModules ++ pathsModule) \\ exposedModules) fromList mOtherModules ++ maybe [] fromList mGeneratedOtherModules
+    generatedModules = maybe [] fromList mGeneratedOtherModules ++ maybe [] fromList mGeneratedExposedModules
 
 fromLibrarySectionInConditional :: [String] -> LibrarySection -> Library
-fromLibrarySectionInConditional inferableModules lib@(LibrarySection _ exposedModules otherModules _ _) = do
+fromLibrarySectionInConditional inferableModules lib@(LibrarySection _ exposedModules _ otherModules librarySectionGeneratedOtherModules _ _) = do
   case (exposedModules, otherModules) of
-    (Nothing, Nothing) -> (fromLibrarySectionPlain lib) {libraryOtherModules = inferableModules}
+    (Nothing, Nothing) -> (fromLibrarySectionPlain lib) {libraryOtherModules = inferableModules ++ fromMaybeList librarySectionGeneratedOtherModules}
     _ -> fromLibrarySectionPlain lib
 
 fromLibrarySectionPlain :: LibrarySection -> Library
 fromLibrarySectionPlain LibrarySection{..} = Library {
     libraryExposed = librarySectionExposed
-  , libraryExposedModules = fromMaybeList librarySectionExposedModules
-  , libraryOtherModules = fromMaybeList librarySectionOtherModules
+  , libraryExposedModules = fromMaybeList librarySectionExposedModules ++ fromMaybeList librarySectionGeneratedExposedModules
+  , libraryOtherModules = fromMaybeList librarySectionOtherModules ++ fromMaybeList librarySectionGeneratedOtherModules
+  , libraryGeneratedModules = fromMaybeList librarySectionGeneratedOtherModules ++ fromMaybeList librarySectionGeneratedExposedModules
   , libraryReexportedModules = fromMaybeList librarySectionReexportedModules
   , librarySignatures = fromMaybeList librarySectionSignatures
   }
@@ -957,9 +964,10 @@ toExecutable dir packageName_ globalOptions =
   where
     fromExecutableSection :: [String] -> [String] -> ExecutableSection -> Executable
     fromExecutableSection pathsModule inferableModules ExecutableSection{..} =
-      (Executable executableSectionMain otherModules)
+      (Executable executableSectionMain (otherModules ++ generatedModules) generatedModules)
       where
         otherModules = maybe (inferableModules ++ pathsModule) fromList executableSectionOtherModules
+        generatedModules = maybe [] fromList executableSectionGeneratedOtherModules
 
 expandMain :: Section ExecutableSection -> Section ExecutableSection
 expandMain = flatten . expand