+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   - `extra-doc-files` requires setting `cabal-version` to at least
     1.18; this is now done properly.
   - Accept bool for `condition` (see #230)
+  - `generated-exposed-modules` and `generated-other-modules`, for populating
+    the `autogen-modules` field (#207).
 
 ## Changes in 0.21.2
   - Fix a bug in module inference for conditionals (see #236)